feat: add bootstrap templates, crispy forms, user profile model

+++ src/blog/views.py
@@ -1,6 +1,7 @@
+from django.http.response import HttpResponse
 from django.shortcuts import get_object_or_404, redirect, render
-from .models import Post
-from .forms import PostForm
+from .models import Post, Like
+from .forms import CommentForm, PostForm
 
 
 def post_list(request):
@@ -28,21 +29,35 @@ def post_create(request):
     return render(request, "blog/post_create.html", context)
 
 def post_detail(request, slug):
+    form = CommentForm()
     obj = get_object_or_404(Post, slug= slug)
+    if request.method == "POST":
+        form = CommentForm(request.POST)
+        if form.is_valid():
+            comment = form.save(commit=False)
+            comment.user = request.user
+            comment.post = obj
+            comment.save()
+            return redirect("blog:detail", slug=slug)
+            # return redirect(request.path)
     context = {
-        "object":obj
+        "object":obj,
+        "form":form
     }
     return render(request, "blog/post_detail.html", context)
 
 def post_update(request, slug):
     obj = get_object_or_404(Post, slug=slug)
     form = PostForm(request.POST or None, request.FILES or None, instance=obj)
+    if request.user.id != obj.author.id:
+        # return HttpResponse("You're not authorized")
+        return redirect("blog:list")
     if form.is_valid():
         form.save()
         return redirect("blog:list")
     
     context = {
-        "objec":obj,
+        "object":obj,
         "form":form
     }
     
@@ -50,6 +65,9 @@ def post_update(request, slug):
 
 def post_delete(request, slug):
     obj = get_object_or_404(Post, slug=slug)
+    if request.user.id != obj.author.id:
+        # return HttpResponse("You're not authorized")
+        return redirect("blog:list")
     if request.POST=="POST":
         obj.delete()
         return redirect("blog:list")
@@ -58,6 +76,16 @@ def post_delete(request, slug):
         "object":obj
     }
     return render(request, "blog/post_delete.html", context)
+
+def like(request, slug):
+    if request.method == "POST":
+        obj = get_object_or_404(Post, slug= slug)
+        like_qs = Like.objects.filter(user=request.user, post=obj)
+        if like_qs.exists():
+            like_qs[0].delete()
+        else:
+            Like.objects.create(user=request.user, post=obj)
+        return redirect("blog:detail", slug=slug)
